Pass connection instances between entities
In order to allow for custom database options to be passed into the
application, the `pqxx::connection` instance could no longer be a
singleton.

Instead, having entities pass a database instance between each other
ensures that they remain decoupled from any sort of global object, while
still being able to share a single database connection.

A database class, `postgresql`, was created to act as a wrapper around
the database connection. Its job is to simply provide a connection
object for entities. This allows for things like connection pools in the
future.