Use string references in place of string views
String views are inefficient when it comes to database operations as all
views must be placed back into string variables. Using string references
allows this extra allocation to be skipped.

The connection type for postgresql is also switched from `connection_base`
to `connection` since `connection_base` is deprecated in libpqxx 7.0.5.