Avoid reading values based on function arguments
Due to the order of execution in C++, the calls to `read<Arg>()` in
`protocol.use` would be executed in a random order. In practice, they
would be called in reverse. This means the values for function
parameters would be read in from the client in reverse order.

The fix for this is to simply read the values manually by interacting
directly with the protocol object.